Move feature request from Type to Triage
I noted which categories are appropriate for PRs and which are appropriate for issues
I added the new Expeditor: Bump Version Major label now that Expeditor can perform major version bumps
I shuffled things around in status and merged the help categories into status. These indicate where things are in resolving an issue. Is something already working on it. Do we want community members to grab it. Is it ideal for a new contributor.
Feature request was moved from type into triage since it's a label we set on issues during triage. I also removed the type of question since we already had that covered under triage as well. If someone opens a question we're going to close it out with a canned response redirecting them to Slack / Discourse.